changed input_units_* to units_*. thru GetMpiTasks Changed GetIPhreeqcId to GetIPhreeqcPointer added to tests RM_GetIPhreeqcId
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@8586 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/pitzer.cpp
@@ -543,7 +543,7 @@ pitzer_tidy(void)
 	/* remake map */
 	{
 		pitz_param_map.clear();
-		for (size_t j = 0; j < count_pitz_param; j++)
+		for (int j = 0; j < count_pitz_param; j++)
 		{	
 			std::set< std::string > header;
 			for (int i = 0; i < 3; i++)
